refactor(server): improve tutors cache logic
