+++ package.json
@@ -110,6 +110,5 @@
       "core-js",
       "protobufjs"
     ]
-  },
-  "packageManager": "pnpm@10.6.1+sha256.81204845a396962a9c4b49cc2d6cafbb49a758652d3c243f2128cbc638084fe2"
+  }
 }
