+++ services/server/src/workers/handlers.ts
@@ -5,7 +5,7 @@ import jwt from "jsonwebtoken";
 import { tokensExpireTime, jwtSecret } from "@/constants";
 import { safe } from "@litespace/utils/error";
 import { nameof } from "@litespace/utils/utils";
-import { WorkerMessagePayloadMap } from "@/workers/types";
+import { WorkerMessageOf } from "@/workers/types";
 import { tutors } from "@litespace/models";
 import { isOnboard, joinTutorCache } from "@/lib/tutor";
 import { cache } from "@/lib/cache";
@@ -46,29 +46,19 @@ export async function sendAuthTokenEmail({
  * This function is responsible of updating the tutors cache data and verifying its integrity
  */
 export async function updateTutorCache(
-  payload: WorkerMessagePayloadMap["update-tutor-in-cache"]
+  payload: WorkerMessageOf<"update-tutor-cache">["payload"]
 ) {
   const error = await safe(async () => {
     await cache.connect();
     const tutor = await tutors.findById(payload.tutorId);
     if (!tutor) return;
 
-    // remove tutors from cache in case they are no longer "onboard"
-    if (!isOnboard(tutor)) {
-      await cache.tutors.removeOne(payload.tutorId);
-      return;
-    }
+    // Remove tutors from cache in case they are no longer "onboard"
+    if (!isOnboard(tutor)) return await cache.tutors.removeOne(payload.tutorId);
 
     const tutorCache = await cache.tutors.getOne(tutor.id);
     const joinedCache = await joinTutorCache(tutor, tutorCache);
     await cache.tutors.setOne(joinedCache);
-
-    /* TODO: Notify clients (users) that a tutor cache data has been updated
-     * So users keep up-to-date
-      context.io
-        .to(Wss.Room.TutorsCache)
-        .emit(Wss.ServerEvent.TutorUpdated, tutor);
-    */
   });
 
   if (error instanceof Error) console.error(error);
