+++ services/server/src/workers/types.ts
@@ -1,33 +1,20 @@
-import { Server } from "@litespace/types";
+export type WorkerMessage =
+  | {
+      type: "send-user-verification-email" | "send-forget-password-email";
+      payload: {
+        user: number;
+        email: string;
+        callbackUrl: string;
+      };
+    }
+  | {
+      type: "update-tutor-cache";
+      payload: { tutorId: number };
+    };
 
-export type WorkerMessagePayloadMap = {
-  "send-user-verification-email": {
-    user: number;
-    email: string;
-    callbackUrl: string;
-  };
-  "send-forget-password-email": {
-    user: number;
-    email: string;
-    callbackUrl: string;
-  };
-  "update-tutor-in-cache": {
-    tutorId: number;
-  };
-};
+export type WorkerMessageType = WorkerMessage["type"];
 
-export type WorkerMessageType = keyof WorkerMessagePayloadMap;
-
-export type WorkerMessage<T extends WorkerMessageType> = {
-  type: T;
-  payload: WorkerMessagePayloadMap[T];
-};
-
-export enum PartentPortMessageType {
-  Stats,
-}
-
-export type PartentPortMessage = {
-  type: PartentPortMessageType.Stats;
-  stats: Server.Stats;
-};
+export type WorkerMessageOf<T extends WorkerMessageType> = Extract<
+  WorkerMessage,
+  { type: T }
+>;
